+++ app/Http/Resources/FeatResource.php
@@ -21,6 +21,12 @@ class FeatResource extends JsonResource
             'prerequisites_text' => $this->prerequisites_text,
             'prerequisites' => EntityPrerequisiteResource::collection($this->whenLoaded('prerequisites')),
             'description' => $this->description,
+
+            // Computed fields
+            'is_half_feat' => $this->is_half_feat,
+            'parent_feat_slug' => $this->parent_feat_slug,
+
+            // Relationships
             'modifiers' => ModifierResource::collection($this->whenLoaded('modifiers')),
             'proficiencies' => ProficiencyResource::collection($this->whenLoaded('proficiencies')),
             'conditions' => EntityConditionResource::collection($this->whenLoaded('conditions')),